Deploy app to S3 for staging
We need to be able to share this app for user testing.

Create a script that renders the output of the express app to a string
and use this in an npm script, along with an S3 CLI tool to deploy the
project to S3.

Addresses https://app.asana.com/0/0/380375711802847/f