Use React
In preparation to implement search, I realized that there was going to
be a lot of UI elements and that the coordination of them would break
under my existing rendering strategy.  Because I'm familiar with React,
it seemed like a reasonable choice to implement views/components.

Get rid of the Nunjucks template and turn it into a bunch of React
components and render these in the Express app.

Update the data transformation flow to prepare initial state server
side, and get rid of data attributes and code that read initial state
from the DOM.  Instead, we inline it as JSON.

Since React is now handling the rendering and events, wrap the
initialization code in index.js in a function instead of a class.

Implement Redux-style container components (see
http://redux.js.org/docs/basics/UsageWithReact.html).

Update the Babel configuration to handle the JSX.

Update eslint configuration to handle React/JSX.