Implement search
Add a presentational component and a container for a search form.

Add a search index to the state using the `js-search` package.

Add an action creator to initiate search.

Rework the reducers so the search and status filters can work together.
Toggle the visibility of recommendations based on a single set of ids.

Restructure the state so that we have a lookup of recommendation objects
and a list of ids.  This allows us to avoid iterating through all of the
recommendations when we're toggling the collapsed state.