Add router
We eventually want to filter by URL and link to a specific
recommendation card. In preparation, add React Router and configure the
router with a single root route.

Note that we're using `HashRouter` since we're serving this page with
WordPress and I can't think of an easy/clean way to support server-side
routing for our single-page app.