Update filter toggling behavior
A few users commented that the filter behavior, where all items were
initially selected, and the initial click turned a filter off, was
confusing.

Change the behavior so no filters are shown as selected and all
recommendations are shown when the application loads.

When the first filter button is clicked, enable filtering and change
the button style to show that it's selected.

Addresses https://app.asana.com/0/0/406067886364905/f