Implement hash-based routing
Make it possible to have a unique URL for each recommendation, for a set
of recommendations filtered to a certain status and for a set of
recommendations filtered to a certain category.

This makes social sharing easier for users and people at the Reporter
doing social media.

Define routes and parameters using React Router.

Pass route parameters from React Router to the container components.

Move filtering logic for recommendations out of the reducers an into the
container component so filters specified both as route params and from
users clicking the filters can be applied.

These updates require a few new utility functions.

Also, we run into some weirdness where we essentially have two sources
of state.  The URL and the state from Redux.  We want the React Router
params to only affect the initial rendering, so we add an `initialRender`
property to the state and pass this to various components and action
creators.

Addresses https://app.asana.com/0/0/406067886364906/f