Hide category filters
To make the user interface (UI) more compact, hide the category
filters in an expanding drawer.

Implement the components for the drawer and the button that toggles
the drawer's open and closed state.

Add a state property for the drawer's state and action creators and
reducers for manipulating it.

Addresses https://app.asana.com/0/0/406442525108037/f