Rework WordPress publishing script
Don't reference system globals in the function that does most of the
work.  Pass objects in the script in `bin` instead.  This makes the
function with the important logic more testable.  I should probably do
this for all the CLI scripts, but this is the one I'm working on at the
moment.