Remove errant comma when joining stdin chunks
I didn't realize that String.join() defaults to a comma as the
separator.  This was causing a comma to be inserted into the JSON
serialization of the application state when publishing which in
turn caused the application interactivity to not work.  Fix this
by specifing an empty string as the separator when calling
`String.join()`, i.e., `chunks.join('')`.