Move "read more" button into gist heading
To make the recommendation card more compact, and to avoid
reordering of elements, that is, moving the properties down when
expanded and keeping the full recommendation adjacent to the gist
text, move the "read more" button so it's inline with the gist text.

This is very similar to the original approach, where the gist label
toggled showing more/less, but I think it's better to have an explicit
call to action button with text than to have an icon.

Addresses https://app.asana.com/0/0/423424635618409/f