Animate the expanding recommendation information
This attempts to make a coherant user experience when the user expands
a recommendation card.  It's important to Jonah, and I agree, that we
keep the recommendation gist and the full recommendation close together.
However, to do this, we have to slide the status, category and date
down, which might be jarring.  Animate the effect so that users can
get more of a sense where those elements are going.

This should be as simple as using a CSS transition, but you can't
transition to/from auto height, so we have to use some JavaScript
to help.

Addresses https://app.asana.com/0/0/423424635618409/f