Inject the sharing URL from an environment variable
Previously, the social sharing components got the URL to use for
social sharing from `window.location`.  However, this caused the
HTML being rendered by the static site to be different than that
rendered dynamically after page load because `window.location` isn't
available when rendering the initial static HTML.  This is because
it's rendered by Webpack and Node.js rather than in the browser.

The consequence of having different initial HTML between the static
render and the initial dynamic React render is not clear to me,
but React issues a warning in the console.

To get around this, define a new setting to set the URL for sharing
and pass this value down to the sharing widget components.  That
way the same URL will be available for both the initial static
rendering and dynamic rendering.