Use HashRouter when rendering static HTML
Use `HashRouter` wehn rendering static HTML so that `<Link>` components
have a `#` in the `href` attribute of the rendered `<a>` tag, matching
the client-side code. Otherwise, we get a warning with React and
we may degrade performance.

When doing this, we also have to fake a DOM for the static site
builder by using jsdom.